modbus-proxy-rs: add patch to run more tolerant with slow clients
Some modbus clients react quite slow and need some time after the
connection was established to settle down and be able to receive and
respond to messages.
This patch adds a hardcoded delay of 1000 msecs after the connection was
established. This specific delay works well with Huawei SUN2000 inverter
which seems to be quite 'special' in terms of modbus communication.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>